requirements.txt
tensorflow==2.10
Flask==3.1.0
pillow==11.1.0
opencv-python==4.11.0.86
matplotlib==3.9.4
pandas==2.2.3
scipy==1.13.1
scikit-learn==1.6.1
dm-sonnet==2.0.2
torch==2.6.0
torchvision==0.21.0
Flask-SQLAlchemy==3.0.3
Flask-Bcrypt==1.0.1
